Allow multiple profiles to use the same data.
The profile file now allows a 'download' string that specifies where to download
and read the data. This allows multiple profiles to be active without having to
download for both of them.

Also, the fetch option will cause the data to be printed after downloading. A
new '-q' option suppresses output to reverse this behavior.